Fix incorrect target_link_libraries() discovered in Meson port
We call functions in vgm-emu, but forgot to include it. It only compiles
in CMake because the headers were somehow visible, and the static
library was included indirectly. It fails to link on Meson. Let's add it
anyway to be safe.

I removed vgm-audio becauase we don't use it (real-time playback).